requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
gymnasium>=0.28.0
d4rl @ git+https://github.com/Farama-Foundation/D4RL.git
mujoco>=2.3.0

# Configuration and logging
pyyaml>=6.0
mlflow>=2.5.0
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
